rename params.go to pool_params.go, consolidate pool functions to pool.go
