+++ x/gamm/pool-models/stableswap/pool.go
@@ -15,6 +15,24 @@ import (
 
 var _ types.PoolI = &Pool{}
 
+// NewStableswapPool returns a stableswap pool
+// Invariants that are assumed to be satisfied and not checked:
+// * len(initialLiquidity) = 2
+// * FutureGovernor is valid
+// * poolID doesn't already exist
+func NewStableswapPool(poolId uint64, stableswapPoolParams PoolParams, initialLiquidity sdk.Coins, futureGovernor string) (Pool, error) {
+	pool := Pool{
+		Address:            types.NewPoolAddress(poolId).String(),
+		Id:                 poolId,
+		PoolParams:         stableswapPoolParams,
+		TotalShares:        sdk.NewCoin(types.GetPoolShareDenom(poolId), types.InitPoolSharesSupply),
+		PoolLiquidity:      initialLiquidity,
+		FuturePoolGovernor: futureGovernor,
+	}
+
+	return pool, nil
+}
+
 func (p Pool) GetAddress() sdk.AccAddress {
 	addr, err := sdk.AccAddressFromBech32(p.Address)
 	if err != nil {
